Simplify things a tad
