+++ packages/alfa-media/src/index.ts
@@ -1,7 +1,1 @@
-export * from "./condition";
-export * from "./feature";
 export * from "./media";
-export * from "./query";
-export * from "./resolver";
-export * from "./type";
-export * from "./value";
